+++ src/linuxArm32Main/kotlin/org/guiVista/gui/widget/button/toggle_button_base.kt
@@ -25,6 +25,7 @@ actual interface ToggleButtonBase : ButtonBase {
     var inconsistent: Boolean
         get() = gtk_toggle_button_get_inconsistent(gtkToggleButtonPtr) == TRUE
         set(value) = gtk_toggle_button_set_inconsistent(gtkToggleButtonPtr, if (value) TRUE else FALSE)
+
     /**
      * Whether the button is displayed as a separate indicator and label. You can call this function on a check button
      * or a radiobutton with `drawIndicator = false` to make the button look like a normal button. This can be used to
@@ -44,12 +45,12 @@ actual interface ToggleButtonBase : ButtonBase {
      * @param userData User data to pass through to the [slot].
      * @return The handler ID for the [slot].
      */
-    fun connectToggledSignal(slot: CPointer<ToggledSlot>, userData: gpointer): ULong =
+    fun connectToggledSignal(slot: CPointer<ToggledSlot>, userData: gpointer): UInt =
         connectGSignal(obj = gtkToggleButtonPtr, signal = TOGGLED_SIGNAL, slot = slot, data = userData)
 
     override fun disconnectSignal(handlerId: ULong) {
         super.disconnectSignal(handlerId)
-        disconnectGSignal(gtkToggleButtonPtr, handlerId)
+        disconnectGSignal(gtkToggleButtonPtr, handlerId.toUInt())
     }
 }
